move to clean code
