+++ tests/system_integrations/test_hpcg_runner.py
@@ -2,8 +2,7 @@ import subprocess
 
 import pytest
 
-from chronus.SystemIntegration.hpcg import HpcgService
-from tests.test_domain.fixtures import mock_subprocess_run
+from chronus.SystemIntegration.application_runners.hpcg import HpcgService
 
 
 @pytest.fixture
