+++ tests/system_integrations/test_sqlite_repository.py
@@ -3,7 +3,7 @@ import pytest
 
 from chronus.domain.Run import Run
 from chronus.domain.system_sample import SystemSample
-from chronus.SystemIntegration.sqlite_repository import SqliteRepository
+from chronus.SystemIntegration.repositories.sqlite_repository import SqliteRepository
 from tests.fixtures import datetime_from_string
